+++ myproject/settings.py
@@ -156,6 +156,7 @@ TEMPLATES = [
                 'core.context_processors.notification_unread_count',
                 'core.context_processors.csp_nonce',
                 'core.context_processors.tilgang',
+                'core.context_processors.versjon',
             ],
         },
     },